analyse_book/requirements.txt
grpcio
googleapis-common-protos
redis
